(clang) fixing clang 18 warnings

+++ src/ksp/ksp.c
@@ -253,8 +253,8 @@ _pgr_ksp(PG_FUNCTION_ARGS) {
             }
         }
 
-        values[0] = Int32GetDatum(funcctx->call_cntr + 1);
-        values[1] = Int32GetDatum(path_id);
+        values[0] = Int32GetDatum((int32_t)funcctx->call_cntr + 1);
+        values[1] = Int32GetDatum((int32_t)path_id);
         values[2] = Int32GetDatum(path[funcctx->call_cntr].seq);
         if (PG_NARGS() != 6) {
             values[3] = Int64GetDatum(path[funcctx->call_cntr].start_id);
